move error functions to a new utils module
